+++ NETMQClient/Program.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using ZeroMQ;
 
 namespace NETMQClient
 {
@@ -11,8 +12,9 @@ namespace NETMQClient
         static void Main(string[] args)
         {
             Console.WriteLine("Enter in NETMQClient");
-            PipeLineSink();
+            MClientMode();
             Console.ReadLine();
         }
+       
     }
 }