+++ NETMQWorker/Program.cs
@@ -4,14 +4,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace NETMQWorker
+namespace NETMQMiddle
 {
     public static partial class Program
     {
         static void Main(string[] args)
         {
-            Console.WriteLine("Enter in NETMQWorker");
-            PipeLineBroker();
+            Console.WriteLine("Enter in NETMQMiddle");
+            MServerMode();
             Console.ReadLine();
         }
     }