+++ History.md
@@ -3,4 +3,5 @@
 =======================
 * Add(): 增加Tcp客户端；增加Tcp服务端
 * Add(): 简单状态机
-* Add(): ZMQ PipeLine 模型: PipeLineVentilator, PipeLineSink, PipeLineBroker
+* Add(): ZMQ PipeLine 模型: PipeLineVentilator, PipeLineSink, PipeLineBroker；应用场景并行处理大数据
+* Add(): ZMQ 多个服务器: MClientMode, MServerMode, MServerMode,可开启多个服务器；应用场景代理与反代理